UI skeleton of NationalFragment.java done
The UI displays a working page viewer with 4 pages. One page for displaying list of braches, one for displaying list of locations, one for displaying durations and last one for showing two cards one for core and the other for non-core. The view pager is fully functional as expected. However, interaction with each list view items is not yet programmed.