UI Done for the OpportunitiesFragment.java and Modified UI of NationalFragment.java pager view
Navigated from main input filters to the next list of opportunities fragment. Here all the opportunities based on the previous filters are displayed (not yet realtime data or filtering done). Also modified the UI of pager view where inputs are taken from user. Instead of having two cards for the core and noncore, now we have two simple buttons that are displayed all the time outside the pager view