UI Done for OpportunityDetailsFragment.java completely
Now, when the user clicks any of the items in the opportunity list, they will be taken to the detailed screen. Detailed screen contains only one pager which contains 2 pages in it. 1st Page shows only the summary, 2nd page laods another pager. This inner pager contains 5 pages, one for job details, description, skills, requirements, and photos each. As of now, the tabs are all working perfectly fine and shows the apt fragment in them.