UI done for Companies and NGOs screens
Complete UI has been done for displaying the list of companies in grid view when the user clicks the companies item in the nav drawer. The items are clickable and when clicked, the apt listener gets the position of item clicked and the item view. Also navigation from home screen to NGO screen is done. The UI is exactly replicated from InternshipsScreen which has two cards for national and international each. Here its NGOs and Skill Dev