UI done for NGOScreen complete flow
Starting from the item in nvaigation drawer, the entire UI navigation flow has been done for the ngos screen. When the user selects ngos card, they are taken to a list of ngos. Each ngo is exapnadable and shows more details when expanded. Address of each ngo is displayed in the form of a custom alert dialog. However, the details of the UI are neglected as of now.