UI done for the skills dev card completely
The flow is from the skills card on ngo screen to making a call from the list. Similar to the ngos card, even this card displays a list of available skills each of which can be expanded or collapsed to see more or less details. Each skill item has a phone number. On clicking the phone number, if there is permission to call, the user can make a call to that number. All the above works with some dummy data as of now. But the flow is flawless.