Register module completed entirely - Back end & Front end
UI was created for the register module. Inputs taken from the user are validated and sent to background service which would connect to DB and get the cknowledgement message back. Based on this, the user is either logged in or declined to login. All the functionlaity seems to be working fine as of now